pyproject.toml
[tool.poetry]
name = "hw-09"
version = "0.1.0"
description = ""
authors = ["Виталий Личинин <lichinin.v@yandex.ru>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
pytest = "^8.2.1"
pylint = "^3.2.2"
requests = "^2.32.2"
selenium = "^4.21.0"
allure-pytest = "^2.13.5"
faker = "^25.3.0"
pytest-xdist = "^3.6.1"
cerberus = "^1.3.5"
pydantic = "^2.7.3"
mysql-connector-python = "^8.4.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
